Keyboard Style Fixed
I fixed App.xaml to work with the new array-indexed buttons. I moved the
KeyButton class to a seperate .cs using it's own namespace and updated
the code in MainWindow.cs to reflect that. I also added a similar style
for the tab item buttons and made it so the currently selected tab
button is dark grey.